pyproject.toml
[project]
name = "rag-system"
version = "0.1.0"
description = "A Rag system"
readme = "README.md"
requires-python = ">=3.11,<3.13"

dependencies = [
  "langchain-community>=0.4.1",
  "llama-index>=0.14.13",
  "chromadb>=1.4.1",
  "flagembedding>=1.3.5",
  "langchain>=1.2.7",
  "langchain-text-splitters>=0.2.0",
  "pydantic-settings>=2.12.0",
  "pydantic-settings>=2.12.0",
  "pymupdf>=1.26.6",
  "pymupdf-layout>=1.26.6",
  "pymupdf4llm>=0.2.9",
  "tesseract>=0.1.3",
  "torch>=2.10.0",
  "transformers==4.44.2",
  "types-pytz>=2025.2.0.20251108",
]

[tool.ruff]
line-length = 120

target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]



[tool.mypy]
plugins = ["pydantic.mypy"]
python_version = "3.11"
strict = true
ignore_missing_imports = true
explicit_package_bases = true
mypy_path = ["src"]

[dependency-groups]
dev = [
  "bump-pydantic>=0.8.0",
  "mypy>=1.19.1",
  "pytest>=9.0.2",
  "ruff>=0.14.14",
  "taplo>=0.9.3",
  "ty>=0.0.15",
  "vulture>=2.14",
]
